Break out the `stack ...` command hierarchy
This commit breaks out the `stack ...` command hierarchy into its own
subpackage, continuing the clean up of `pkg/cmd/pulumi`.
