+++ pkg/cmd/pulumi/stack/stack_import.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2022, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,29 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package stack
 
 import (
-	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"os"
 
-	"github.com/hashicorp/go-multierror"
-
 	"github.com/spf13/cobra"
 
-	"github.com/pulumi/pulumi/pkg/v3/backend"
 	"github.com/pulumi/pulumi/pkg/v3/backend/display"
 	cmdBackend "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/backend"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/cmd"
-	cmdStack "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/stack"
-	"github.com/pulumi/pulumi/pkg/v3/resource/deploy"
 	"github.com/pulumi/pulumi/pkg/v3/resource/stack"
 	pkgWorkspace "github.com/pulumi/pulumi/pkg/v3/workspace"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/apitype"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/diag"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/cmdutil"
 )
 
@@ -60,12 +52,12 @@ func newStackImportCmd() *cobra.Command {
 			}
 
 			// Fetch the current stack and import a deployment.
-			s, err := cmdStack.RequireStack(
+			s, err := RequireStack(
 				ctx,
 				ws,
 				cmdBackend.DefaultLoginManager,
 				stackName,
-				cmdStack.LoadOnly,
+				LoadOnly,
 				opts,
 			)
 			if err != nil {
@@ -95,7 +87,7 @@ func newStackImportCmd() *cobra.Command {
 			if err != nil {
 				return checkDeploymentVersionError(err, s.Ref().Name().String())
 			}
-			if err := saveSnapshot(ctx, s, snapshot, force); err != nil {
+			if err := SaveSnapshot(ctx, s, snapshot, force); err != nil {
 				return err
 			}
 			fmt.Printf("Import complete.\n")
@@ -112,71 +104,4 @@ func newStackImportCmd() *cobra.Command {
 		&file, "file", "", "", "A filename to read stack input from")
 
 	return cmd
-}
-
-func saveSnapshot(ctx context.Context, s backend.Stack, snapshot *deploy.Snapshot, force bool) error {
-	stackName := s.Ref().Name()
-	var result error
-	for _, res := range snapshot.Resources {
-		if res.URN.Stack() != stackName.Q() {
-			msg := fmt.Sprintf("resource '%s' is from a different stack (%s != %s)",
-				res.URN, res.URN.Stack(), stackName)
-			if force {
-				// If --force was passed, just issue a warning and proceed anyway.
-				// Note: we could associate this diagnostic with the resource URN
-				// we have.  However, this sort of message seems to be better as
-				// something associated with the stack as a whole.
-				cmdutil.Diag().Warningf(diag.Message("" /*urn*/, msg))
-			} else {
-				// Otherwise, gather up an error so that we can quit before doing damage.
-				result = multierror.Append(result, errors.New(msg))
-			}
-		}
-	}
-	// Validate the stack. If --force was passed, issue an error if validation fails. Otherwise, issue a warning.
-	if !backend.DisableIntegrityChecking {
-		if err := snapshot.VerifyIntegrity(); err != nil {
-			msg := fmt.Sprintf("state file contains errors: %v", err)
-			if force {
-				cmdutil.Diag().Warningf(diag.Message("", msg))
-			} else {
-				result = multierror.Append(result, errors.New(msg))
-			}
-		}
-	}
-	if result != nil {
-		return multierror.Append(result,
-			errors.New("importing this file could be dangerous; rerun with --force to proceed anyway"))
-	}
-
-	// Explicitly clear-out any pending operations.
-	if snapshot.PendingOperations != nil {
-		for _, op := range snapshot.PendingOperations {
-			msg := fmt.Sprintf(
-				"removing pending operation '%s' on '%s' from snapshot", op.Type, op.Resource.URN)
-			cmdutil.Diag().Warningf(diag.Message(op.Resource.URN, msg))
-		}
-
-		snapshot.PendingOperations = nil
-	}
-	sdp, err := stack.SerializeDeployment(ctx, snapshot, false /* showSecrets */)
-	if err != nil {
-		return fmt.Errorf("constructing deployment for upload: %w", err)
-	}
-
-	bytes, err := json.Marshal(sdp)
-	if err != nil {
-		return err
-	}
-
-	dep := apitype.UntypedDeployment{
-		Version:    apitype.DeploymentSchemaVersionCurrent,
-		Deployment: bytes,
-	}
-
-	// Now perform the deployment.
-	if err = s.ImportDeployment(ctx, &dep); err != nil {
-		return fmt.Errorf("could not import deployment: %w", err)
-	}
-	return nil
 }
