+++ pkg/cmd/pulumi/stack/stack_output.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package stack
 
 import (
 	"context"
@@ -31,7 +31,6 @@ import (
 	"github.com/pulumi/pulumi/pkg/v3/backend/display"
 	cmdBackend "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/backend"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/cmd"
-	cmdStack "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/stack"
 	"github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/ui"
 	"github.com/pulumi/pulumi/pkg/v3/resource/deploy"
 	"github.com/pulumi/pulumi/pkg/v3/resource/stack"
@@ -82,7 +81,7 @@ type stackOutputCmd struct {
 	// from tests.
 	requireStack func(
 		ctx context.Context, ws pkgWorkspace.Context, lm cmdBackend.LoginManager,
-		name string, lopt cmdStack.LoadOption, opts display.Options,
+		name string, lopt LoadOption, opts display.Options,
 	) (backend.Stack, error)
 
 	Stdout io.Writer // defaults to os.Stdout
@@ -93,7 +92,7 @@ func (cmd *stackOutputCmd) Run(ctx context.Context, args []string) error {
 		Color: cmdutil.GetGlobalColorization(),
 	}
 
-	requireStack := cmdStack.RequireStack
+	requireStack := RequireStack
 	if cmd.requireStack != nil {
 		requireStack = cmd.requireStack
 	}
@@ -129,7 +128,7 @@ func (cmd *stackOutputCmd) Run(ctx context.Context, args []string) error {
 		cmd.ws,
 		cmdBackend.DefaultLoginManager,
 		cmd.stackName,
-		cmdStack.LoadOnly,
+		LoadOnly,
 		opts,
 	)
 	if err != nil {
@@ -166,7 +165,7 @@ func (cmd *stackOutputCmd) Run(ctx context.Context, args []string) error {
 	}
 
 	if cmd.showSecrets {
-		log3rdPartySecretsProviderDecryptionEvent(ctx, s, "", "pulumi stack output")
+		Log3rdPartySecretsProviderDecryptionEvent(ctx, s, "", "pulumi stack output")
 	}
 
 	return nil
