+++ pkg/cmd/pulumi/stack/stack_output_fuzz_test.go
@@ -15,7 +15,7 @@
 //go:build go1.18
 // +build go1.18
 
-package main
+package stack
 
 import (
 	"bytes"
