+++ pkg/cmd/pulumi/stack/stack_test.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2013, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package stack
 
 import (
 	"bytes"
@@ -20,6 +20,7 @@ import (
 	"testing"
 
 	"github.com/pulumi/pulumi/pkg/v3/backend"
+	cmdBackend "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/backend"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/tokens"
 	"github.com/stretchr/testify/assert"
 )
@@ -120,4 +121,12 @@ func TestStringifyOutput(t *testing.T) {
 			assert.Equal(t, tt.want, got)
 		})
 	}
+}
+
+// mockBackendInstance sets the backend instance for the test and cleans it up after.
+func mockBackendInstance(t *testing.T, b backend.Backend) {
+	t.Cleanup(func() {
+		cmdBackend.BackendInstance = nil
+	})
+	cmdBackend.BackendInstance = b
 }
